+++ library/core/src/main/java/com/google/android/exoplayer2/Player.java
@@ -89,7 +89,7 @@ public interface Player {
      * default audio attributes will be used. They are suitable for general media playback.
      *
      * <p>Setting the audio attributes during playback may introduce a short gap in audio output as
-     * the audio track is recreated. A new audio session id will also be generated.
+     * the audio track is recreated.
      *
      * <p>If tunneling is enabled by the track selector, the specified audio attributes will be
      * ignored, but they will take effect if audio is later played without tunneling.